Conflicts:
	AppSalesMobile.xcodeproj/project.pbxproj
	Classes/App.h
	Classes/App.m
	Classes/AppCell.m
	Classes/AppSalesMobileAppDelegate.h
	Classes/AppSalesMobileAppDelegate.m
	Classes/CountriesController.m
	Classes/Country.h
	Classes/Country.m
	Classes/Day.h
	Classes/Day.m
	Classes/DayCell.m
	Classes/DaysController.m
	Classes/Entry.m
	Classes/ReportManager.h
	Classes/ReportManager.m
	Classes/Review.h
	Classes/Review.m
	Classes/ReviewsController.m
	Classes/ReviewsListController.h
	Classes/ReviewsListController.m
	Classes/RootViewController.m
	Classes/SettingsViewController.m
	Classes/TrendGraphView.m
	Classes/WeekCell.m
	Classes/WeeksController.m
	English.lproj/SettingsViewController.xib
	German.lproj/Localizable.strings
	Images/Default.png
	Info.plist
	README.textile
	ReviewTemplate.html

+++ Classes/ReviewCell.m
@@ -78,7 +78,7 @@
 	[[review.countryCode uppercaseString] drawInRect:CGRectMake(0, 53, 44, 9) withFont:[UIFont systemFontOfSize:9.0] lineBreakMode:UILineBreakModeWordWrap alignment:UITextAlignmentCenter];
 	
 	NSMutableString *starsString = [NSMutableString string];
-	for (int i = 0; i < review.stars; i++) {
+	for (NSUInteger i = 0; i < review.stars; i++) {
 		[starsString appendString:@"★"];
 	}
